Fix manifest for case-sensitive filesystems
The file is named `background/keyBoardControls.js` whereas the manifest
refers it as `background/keyboardControls.js`. This is fine for
case-insensitive filesystems such as NTFS (Windows) or default HFS
(MacOS), but the installation of the plugin will fail on case-sensitive
filesystems, which is the default on Linux.

This commit fixes the issue by changing the case in the manifest.